+++ utils/loghandlers.py
@@ -19,7 +19,49 @@
 # $Id: loghandlers.py 301527 2017-02-06 15:31:43Z matj $
 #
 
-import logging
+import logging, os
+
+_artifactsLog = logging.getLogger('xpybuild.artifacts')
+# keep track of formatters that have been instantiated to allow us to publish artifacts to them
+_outputFormattersInUse = []
+def publishArtifact(displayName, path):
+	""" Publishes the specified local path as an artifact, 
+	if supported by the configured output format. 
+	
+	For example this can be used to publish log and error output if a target 
+	fails. 
+	
+	Note that many output formatters do not have any concept of artifact 
+	publishing, so it is usually desirable to log important paths at INFO 
+	or WARN/ERROR or above in addition to calling this method. 
+	
+	@param displayName: A string describing the artifact being published. 
+	Some output formatters may ignore this string. 
+	
+	@param path: An absolute path to a file (e.g. a log file) that should 
+	be published, if supported. Empty string or None are ignored. 
+	"""
+	if not path:
+		_artifactsLog.debug('Ignoring empty artifact path: "%r"', path)
+		return
+	
+	assert displayName, 'displayName must be specified'
+	
+	# we do error checking at this level so we can detect errors regardless 
+	# of what output formatter is configured
+	
+	if not path or not os.path.isabs(path):
+		raise Exception('Cannot publish artifact path "%s" because only absolute paths are supported'%path)
+
+	path = os.path.normpath(path)
+	
+	if not os.path.exists(path):
+		_artifactsLog.warning("Cannot find path for artifact publishing: \"%s\"", path)
+	# always log at debug, so they're there 
+	_artifactsLog.debug('Publishable artifact path for %s: "%s"', displayName, path)
+
+	for f in _outputFormattersInUse:
+		f.publishArtifact(_artifactsLog, displayName, path)
 
 class LogHandler(object):
 	"""
@@ -35,14 +77,36 @@ class LogHandler(object):
 	def __init__(self):
 		self.fmt = logging.Formatter()
 		self.bufferingDisabled = False # can be set to True to prevent it for handlers for which it's not appropriate
-
+		_outputFormattersInUse.append(self)
 		
 	def setLevel(self, level):
 		self.level = level
 	def handle(self, record):
 		raise "Not Implemented"
+	
+	def publishArtifact(self, logger, displayName, path):
+		""" Publishes the specified local path (e.g. a log file) as an artifact, 
+		if supported by this formatter. 
+		
+		The default implementation used by most formatters is a no-op. 
+
+		@param logger: The logger instance that should be used to write any 
+		messages to stdout as part of the publishing process (this will not be
+		relevant for all formatters). 
+		
+		@param displayName: A string describing the artifact being published. 
+		Some output formatters may ignore this string. 
+		
+		@param path: An absolute path to a file (e.g. a log file) that should 
+		be published. Never equal to empty or None. 
+		"""
+		pass
+		
 
 class XpybuildHandler(LogHandler):
+	"""
+	The default text output formatter for xpybuild. 
+	"""
 	def __init__(self, stream, buildOptions):
 		LogHandler.__init__(self)
 		self.delegate = logging.StreamHandler(stream)
@@ -56,6 +120,9 @@ class XpybuildHandler(LogHandler):
 _handlers = {}
 
 def registerHandler(name, handler):
+	"""
+	Called to make a custom output formatter class available for use by xpybuild. 
+	"""
 	_handlers[name] = handler
 
 registerHandler("xpybuild", XpybuildHandler)